% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.clust.index.R
\name{extract.clust.index}
\alias{extract.clust.index}
\title{extract.clust.index.R}
\usage{
extract.clust.index(index, mixmodout)
}
\arguments{
\item{index}{The index of the marker to be extracted}

\item{mixmodout}{Rmixmod results}
}
\value{
A list of Rmixmod cluster statistics for a particular marker
}
\description{
extracts the MixmodResults for a given marker index
}
\examples{


}
