% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.clust.stats.R
\name{extract.clust.stats}
\alias{extract.clust.stats}
\title{extract.clust.stats}
\usage{
extract.clust.stats(indices, GSdata, mixmodout)
}
\arguments{
\item{indices}{A vector of marker indices to extract}

\item{GSdata}{Processed GenomeStudio data}

\item{mixmodout}{Clustering output from step 2}
}
\value{
A
}
\description{
Creates a data.frame of cluster statistics for a set of marker indices
}
\examples{

}
