% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.genotype.R
\name{extract.genotype}
\alias{extract.genotype}
\title{Generates a c(0,1,2) genotype matrix from a list of marker names}
\usage{
extract.genotype(markerlist, GSdata, mixmodout, clust_stats)
}
\arguments{
\item{markerlist}{A character vector of marker names}

\item{GSdata}{A list of R and Theta data.frames}

\item{mixmodout}{Rmixmod clustering results}
}
\value{
A numerical matrix with rows = markers and columns = samples
}
\description{
Uses a gbm model to predict whether to keep or remove clusters
}
\examples{

}
