% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.marker.index.R
\name{extract.marker.index}
\alias{extract.marker.index}
\title{extract.marker.index}
\usage{
extract.marker.index(index, GSdata)
}
\arguments{
\item{index}{The index of the marker to be extracted}

\item{GSdata}{Processed GenomeStudio data}
}
\value{
A list containing only marker name, Theta and R values for the specified marker
}
\description{
extracts the theta and r coordinates for a given marker index
}
\examples{


}
