% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.clusters.R
\name{filter.clusters}
\alias{filter.clusters}
\title{Classify unwanted clusters, remove bad clusters, and output partition table}
\usage{
filter.clusters(GSdata, mixmodout, model_gbm)
}
\arguments{
\item{filepath}{the path to the completed .csv file}
}
\value{
A gbm model object
}
\description{
Uses a gbm model to predict whether to keep or remove clusters
}
\examples{

}
