% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.predict.R
\name{gbm.predict}
\alias{gbm.predict}
\title{Train noise filter}
\usage{
gbm.predict(model_gbm, clust_stats)
}
\arguments{
\item{model_gbm}{A gradient boosted model object created using manually scored training data}

\item{clust_stats}{A cluster statistic object from the extract.clust.stats function}
}
\value{
A gbm model object
}
\description{
Reads in completed training data and trains a gradient boosted model outputting prediction accuracy statistics
}
\examples{

}
