% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.training.data.R
\name{make.training.data}
\alias{make.training.data}
\title{make.training.data}
\usage{
make.training.data(GSdata, mixmodout, marknum = 100, seed = NULL)
}
\arguments{
\item{GSdata}{Processed GenomeStudio data}

\item{mixmodout}{Clustering output from step 2}

\item{marknum}{The number of markers to sample for each cluster category (total = 4 * marknum)}

\item{seed}{Fix the random seed for reproducibility}
}
\value{
A folder containing a partially populated .csv file and a nested folder of marker plots corresponding to the .csv file
}
\description{
Creates a set of training data for manual scoring using excel from a processed Genome Studio export and an Rmixmod cluster output
}
\examples{

}
