% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.raw.clust.R
\name{plot.raw.clust}
\alias{plot.raw.clust}
\title{extract.clust.index}
\usage{
plot.raw.clust(index, GSdata, mixmodout)
}
\arguments{
\item{index}{The index of the marker to be extracted}

\item{GSdata}{Processed GenomeStudio data, 2 item list of 'theta' and 'r' data frames}

\item{mixmodout}{Rmixmod results}
}
\value{
A ggplot scatterplot. Colors indicate cluster membership. Red lines indicate cluster center and relative rotation and variance.
}
\description{
partitions a marker and plots points colored by clusters and surrounded by variance representative elipsoids and cluster centers
}
\examples{


}
