% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGSexport.R
\name{processGSexport}
\alias{processGSexport}
\title{Process Genome Studio export}
\usage{
processGSexport(exportpath = NULL, transform = "arcsin")
}
\arguments{
\item{exportpath}{the path to the GS data export}

\item{transform}{How to transform theta to control variance. (no transformation = 'none', logit transformation = 'logit', arcsine transformation (default) = 'arcsin')}
}
\value{
A list of two data.frames. The first is a table of theta values and the second is a table of intensity values
}
\description{
Processes a Genome Studio data export table to only import useful information for clustering (theta and R values) and ignores the rest
}
