% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.kamiak.scripts.R
\name{SLURM.clust.par}
\alias{SLURM.clust.par}
\title{SLURMclustpar}
\usage{
SLURM.clust.par(GSdata, partition, account, nodes, cpus, job)
}
\arguments{
\item{GSdata}{Processed GenomeStudio data ready for clustering}

\item{partition}{The partition to use on Kamiak}

\item{account}{The account to use on Kamiak}

\item{nodes}{The number of nodes to use}

\item{cpus}{The number of cpus per node to use}

\item{job}{The job ID to use for file creation}
}
\value{
A folder containing the SLURM script and sub-directories containing tables for clustering.
}
\description{
Create a single SLURM script for submitting parallel jobs across multiple Kamiak nodes
}
