% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.noise.filter.R
\name{train.noise.filter}
\alias{train.noise.filter}
\title{Train noise filter}
\usage{
train.noise.filter(trainingpath = "./training.data/training.csv",
  seed = NULL)
}
\arguments{
\item{filepath}{the path to the completed .csv file}
}
\value{
A gbm model object
}
\description{
Reads in completed training data and trains a gradient boosted model outputting prediction accuracy statistics
}
