% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.markers.R
\name{cluster.markers}
\alias{cluster.markers}
\title{cluster.markers}
\usage{
cluster.markers(GSdata)
}
\arguments{
\item{GSdata}{Processed GenomeStudio data, 2 item list of 'theta' and 'r' data frames}
}
\value{
An Rmixmod object containing the best cluster model results
}
\description{
Sets up a parallel backend and clusters each marker using mixture of gaussian models provided by the 'Rmixmod' package.
}
\details{
cluster.markers
}
