% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGSexport.R
\name{processGSexport}
\alias{processGSexport}
\title{Process Genome Studio Export}
\usage{
processGSexport(exportpath = NULL)
}
\arguments{
\item{exportpath}{the path to the GS data export}
}
\value{
A list of two data.frames. The first is a table of theta values and the second is a table of intensity values
}
\description{
Processes a Genome Studio data export table to only import useful information (theta and R values) and ignores the rest
}
